# apps/backend/cmd/docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.Response-any:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      status:
        type: integer
    type: object
...
  controller.Response-service_AuthResponse:
    properties:
      data:
        $ref: '#/definitions/service.AuthResponse'
      message:
        type: string
      meta: {}
      status:
        type: integer
    type: object
...
  controller.Response-service_ProductResponse:
    properties:
      data:
        $ref: '#/definitions/service.ProductResponse'
      message:
        type: string
      meta: {}
      status:
        type: integer
    type: object
...
  controller.ResponsePagination-array_service_ProductResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/service.ProductResponse'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/helpers.Pagination'
      status:
        type: integer
    type: object
...
  dto.LoginDto:
    properties:
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  dto.UserRegisterDto:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - full_name
    - password
    - username
    type: object
  dto.VerifyEmailDto:
    properties:
      token:
        type: string
    type: object
  helpers.Pagination:
    properties:
      count:
        type: integer
      limit:
        type: integer
      next_page:
        type: integer
      page:
        type: integer
      previous_page:
        type: integer
      total_page:
        type: integer
    type: object
  service.AuthResponse:
    properties:
      refresh_token:
        type: string
      token:
        type: string
    type: object
...
  service.CreateProductRequest:
    properties:
      category_id:
        items:
          type: integer
        type: array
      code:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      uom:
        type: string
    required:
    - name
    - price
    - uom
    type: object
...
  service.Image:
    properties:
      created_at:
        type: string
      height:
        type: integer
      id:
        type: integer
      path:
        type: string
      size:
        type: number
      type:
        type: string
      updated_at:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
  service.ProductCategoryResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  service.ProductImage:
    properties:
      created_at:
        type: string
      id:
        type: integer
      image:
        $ref: '#/definitions/service.Image'
      image_id:
        type: integer
      product_id:
        type: integer
      updated_at:
        type: string
    type: object
  service.ProductProductCategoryResponse:
    properties:
      category:
        $ref: '#/definitions/service.ProductCategoryResponse'
      category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      updated_at:
        type: string
    type: object
  service.ProductResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      product_image:
        items:
          $ref: '#/definitions/service.ProductImage'
        type: array
      product_product_category:
        items:
          $ref: '#/definitions/service.ProductProductCategoryResponse'
        type: array
      team_id:
        type: integer
      uom:
        type: string
      updated_at:
        type: string
    type: object
...
  service.UpdateProductRequest:
    properties:
      category_id:
        items:
          type: integer
        type: array
      code:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      uom:
        type: string
    required:
    - code
    - name
    - price
    - uom
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server caller server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /auth/email/send-verify:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - auth
  /auth/email/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailDto'
      produces:
      - application/json
      responses:
        "301":
          description: Moved Permanently
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response-service_AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response-service_AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegisterDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response-service_AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - auth
...
  /teams/{team_id}/products:
    get:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Code
        in: query
        name: code
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: UOM
        in: query
        name: uom
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponsePagination-array_service_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - products
    post:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: CreateProductRequest
        in: body
        name: data
        schema:
          $ref: '#/definitions/service.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response-service_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - products
  /teams/{team_id}/products/{products_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Products ID
        in: path
        name: products_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - products
    get:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Products ID
        in: path
        name: products_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response-service_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - products
    post:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Products ID
        in: path
        name: products_id
        required: true
        type: integer
      - description: UpdateProductRequest
        in: body
        name: data
        schema:
          $ref: '#/definitions/service.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - products
  /teams/{team_id}/products/{products_id}/upload_image:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Products ID
        in: path
        name: products_id
        required: true
        type: integer
      - description: Profile image file
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response-service_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"